Move letter list to own component, add letter count bar component
